Added no-bio option for personnel sections
Use the class 'no-bio' on a .personnel section if it is to be excluded from the bio overlay. This corrects an issue where the hash was displaying incorrectly when a bio overlay was invoked.

Note: if the actual intent here was to not have a direct bio link to the featured consultant (e.g. Michael Scott), but still have the bio available in the overlay carousel, then we will need a separate solution; the assumption with the above fix is that no bio is intended to be surfaced for the featured consultant.