Do not clone forms for use in overlays
This restores the overlay markup to the hidden form container; this retains any added event handlers, which seemed to be the issue as described in regards to SalesForce's implementation.

This likely could also be fixed with a deep clone (.clone(true)), but this approach seems safer given the lack of opportunity to test.